IA-539 refactor: remove redirections from store
